拆分组件
